feat: :sparkles: abstract App fetching logic to useTodayForecastData
