+++ apps/frontend/src/App.tsx
@@ -1,51 +1,16 @@
-import { useState } from 'react';
-import { useQuery } from '@tanstack/react-query';
 import WeatherOverview from './components/WeatherOverview';
-import axios from 'axios';
-import { API_KEY, API_URL } from './constants/api';
-
-export type CityName = 'Gliwice' | 'Hamburg';
-
-interface WeatherData {
-  location: {
-    name: CityName;
-    country: string;
-    localtime: string;
-  };
-  current: {
-    temp_c: string;
-    condition: {
-      text: string;
-      icon: string;
-    };
-  };
-}
-
-const fetchWeatherData = async (city: CityName): Promise<WeatherData> => {
-  const res = await axios.get(`${API_URL}current.json`, {
-    params: {
-      key: API_KEY,
-      q: city,
-      aqi: 'no',
-    },
-  });
-  return res.data;
-};
+import { useTodayForecastData } from './hooks/useTodayForecastData';
 
 function App(): JSX.Element {
-  const [currentCity, setCurrentCity] = useState<CityName>('Gliwice');
-
-  const { isLoading, isError, data, error } = useQuery({
-    queryKey: [currentCity],
-    queryFn: () => fetchWeatherData(currentCity),
-  });
+  const { data, error, isError, isLoading, toggleCity } =
+    useTodayForecastData('Gliwice');
 
   if (isLoading) {
     return <div>Loading...</div>;
   }
 
   if (isError) {
-    return <div>{error.message || 'An unexpected error occurred'}</div>;
+    return <div>{error?.message || 'An unexpected error occurred'}</div>;
   }
 
   if (!data) {
@@ -60,7 +25,7 @@ function App(): JSX.Element {
         localtime={data.location.localtime}
         temperature={data.current.temp_c}
         condition={data.current.condition}
-        setCurrentCity={setCurrentCity}
+        toggleCity={toggleCity}
       />
     </main>
   );
